new implementations on user and post method

+++ src/main/resources/db/migration/V1_create-table-user.sql
@@ -1,6 +1,6 @@
 create table users(
 
-    id bigint not null auto_increment,
+    id serial not null auto_increment,
     email varchar(100) not null unique,
     username varchar(100) not null unique,
     password varchar(255) not null,
